[fix] property setter for idx
