+++ frappe/patches.txt
@@ -81,3 +81,4 @@ execute:frappe.db.sql("update `tabComment` set comment_type='Comment' where comm
 frappe.patches.v5_0.modify_session
 frappe.patches.v5_0.expire_old_scheduler_logs
 execute:frappe.permissions.reset_perms("DocType")
+execute:frappe.db.sql("delete from `tabProperty Setter` where `property` = 'idx'")
