+++ frappe/patches/v5_0/rename_table_fieldnames.py
@@ -2,7 +2,7 @@
 # License: GNU General Public License v3. See license.txt
 
 import frappe
-from frappe.model import rename_field
+from frappe.model.utils import rename_field
 from frappe.modules import scrub, get_doctype_module
 
 rename_map = {
